Changes to abbreviations
Abbreviations are now saved with a graph. For graphs saved without
abbreviation information, the abbreviations can be automatically
generated. However, this feature is not currently implemented, so
saved graphs without `kind_to_abbrev` with raise a NotImplementedError.

Automatic abbreviation generation code has been removed from graph.MetaGraph()
because it was complex/lengthy and not the focus of the graph module.

Currently, kind_to_abbrev is not checked for safety -- i.e. the user could
provide ambiguous abbreviations without an error.